% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biscut_summary_plots.R
\name{biscut_summary_plots}
\alias{biscut_summary_plots}
\title{Produce peak plots}
\usage{
biscut_summary_plots(biscut_results, plot_dir = NULL)
}
\arguments{
\item{biscut_results}{The data (list) that is output from do_biscut()}

\item{plot_dir}{Path for a new directory to save plot PDFs (or NULL to not make any).}
}
\value{
A list containing all the summary plots.
}
\description{
Visualize significant peaks
}
