% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_breakpoints.R
\name{calculate_breakpoints}
\alias{calculate_breakpoints}
\title{Calculate breakpoints for BISCUT}
\usage{
calculate_breakpoints(
  segment_file = NULL,
  output_dir = NULL,
  chromosome_coordinates = get_chromosome_coordinates(),
  arms = get_chromosome_arms(),
  amplitude_threshold = 0.2,
  cores = 1
)
}
\arguments{
\item{segment_file}{Copy number segmentation file path. See README for formatting details.}

\item{output_dir}{Optionally, write breakpoint files to a specified output directory. When NULL
(default), no files will be written.}

\item{chromosome_coordinates}{The coordinates of p and q arms of the human chromosomes. By
default, hg19-based coordinates are used. Please note that these coordinates are not
biologically well-defined regions but rather what appeared to be the boundaries of regions that
can be considered telomeric and (particularly) centromeric from SNP-array based TCGA copy
number profiles. You might want to use your custom set of coordinates based on your copy number
profiles (ex. IGV view of start/end of event near telomere/centromere). From our experience,
choosing the telomeric/centromeric boundaries wide enough so that the events are not missed due
to imperfections of assaying technologies is more important than genome reference differences
(ex., hg19 vs. hg38).}

\item{arms}{Which chromosome arms to process. Defaults to all arms (see ?get_chromosome_arms()). You
may choose to run a subset of arms for testing purposes.}

\item{amplitude_threshold}{Copy number changes less than this threshold are considered noise and
ignored. By default is 0.2 for impurity-corrected input (ISAR correction). We suggest using a
smaller threshold (ex., 0.1) if the input is not corrected for impurity.}

\item{cores}{How many cores to use for processing chromosome arms in parallel. As always with the
\code{parallel} library, use of multiple cores is not supported on Windows systems.}
}
\value{
When successful, a list of two breakpoint data.frames (telomeric and centromeric);
otherwise, attempts to return error information.
}
\description{
Calculate breakpoints for BISCUT
}
