% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BISCUT_peak_finding.R
\name{do_biscut}
\alias{do_biscut}
\title{Run BISCUT analysis on processed arm breakpoint files}
\usage{
do_biscut(
  breakpoint_file_dir,
  results_dir,
  use_precalculated_background = FALSE,
  telcent_thres = 0.001,
  ci = 0.95,
  n_bootstrap = 1000,
  qval_thres = 0.05,
  cores = 1,
  arms = get_chromosome_arms(),
  chromosome_coordinates = get_chromosome_coordinates(),
  gene_locations = get_gene_locations(),
  seed = NULL
)
}
\arguments{
\item{breakpoint_file_dir}{Directory holding breakpoint files as generated by
make_breakpoint_files().}

\item{results_dir}{The path for a results directory that will be created on running.}

\item{use_precalculated_background}{TRUE/FALSE (default FALSE) on whether to use a precalculated
background distribution from TCGA samples. When FALSE, a background distribution is calculated
from input data.}

\item{telcent_thres}{Threshold to filter out very short or whole-arm events to correct for noise
in calling copy number breakpoints. Events that cover a fraction of an arm that is less than
\code{telcent_thres} or greater than \code{(1-telcent_thres)} are removed. Increase the
threshold if you have peaks called too close to the telomere/centromere that you think are not
real.}

\item{ci}{Confidence interval threshold around peaks (larger values will yield wider peaks).}

\item{n_bootstrap}{Number of bootstrapped samples for calculating confidence intervals}

\item{qval_thres}{The threshold on q-values computed for the significance of identified peaks
after correcting for multiple hypothesis testing}

\item{cores}{How many cores to use for processing chromosome arms in parallel. As always with the
\code{parallel} library, use of multiple cores is not supported on Windows systems.}

\item{arms}{Which chromosome arms to analyze. By default, all arms.}

\item{chromosome_coordinates}{The coordinates of p and q arms of the human chromosomes. By
default, hg19-based coordinates are used. Please note that these coordinates are not
biologically well-defined regions but rather what appeared to be the boundaries of regions that
can be considered telomeric and (particularly) centromeric from SNP-array based TCGA copy
number profiles. You might want to use your custom set of coordinates based on your copy number
profiles (ex. IGV view of start/end of event near telomere/centromere). From our experience,
choosing the telomeric/centromeric boundaries wide enough so that the events are not missed due
to imperfections of assaying technologies is more important than genome reference differences
(ex., hg19 vs. hg38).}

\item{gene_locations}{Gene locations for mapping genes to peaks in the output. By default, an hg19-based
table is used; see \code{get_gene_locations()}.}

\item{seed}{Default NULL; supply an integer seed to make numerical output reproducible.}
}
\value{
A list containing BISCUT output:
\itemize{
\item peak_info: A table with one row describing each significant peak.
\item genes_by_peak: A table identifying which genes fall in each peak (and the coordinates of the genes).
\item background_fits: A list containing fitted background length distributions for each event type (centromeric/telomeric amplifications/deletions).
\item peak_plot_data: The data used to make the peak plots (saved to BISCUT results directory).
}
}
\description{
Run BISCUT analysis on processed arm breakpoint files
}
