% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gene_results.R
\name{extract_gene_results}
\alias{extract_gene_results}
\title{Wrapper for filter_BISCUT_knowngenes (Python)}
\usage{
extract_gene_results(results_dir, genes)
}
\arguments{
\item{results_dir}{Output directory from do_biscut().}

\item{genes}{Vector of gene names}
}
\description{
Wrapper for filter_BISCUT_knowngenes (Python)
}
