% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_chromosome_arms}
\alias{get_chromosome_arms}
\title{Get chromosome arms}
\usage{
get_chromosome_arms()
}
\description{
Returns a character vector of all chromosome arms. Acrocentromeric chromosomes are represented by
just chromosome names (e.g., "14"), while the arms of other chromosomes get suffixes (e.g., "2p",
"2q").
}
