% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_breakpoint_files}
\alias{load_breakpoint_files}
\title{Read in breakpoint files from a directory}
\usage{
load_breakpoint_files(breakpoint_file_dir, telcent_thres = 0.001)
}
\arguments{
\item{breakpoint_file_dir}{directory containing breakpoint files}

\item{telcent_thres}{Threshold for filtering out events that are very small or vary large. Events
that cover a fraction of the arm that is less than telcent_thres or greater than
(1-telcent_thres) are removed.}
}
\description{
Read in breakpoint files from a directory
}
