% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peaks.R
\name{plot_peaks}
\alias{plot_peaks}
\title{Produce peak plots}
\usage{
plot_peaks(biscut_results, plot_dir = NULL)
}
\arguments{
\item{biscut_results}{The data (list) that is output from do_biscut()}

\item{plot_dir}{Path for a new directory to save plot PDFs (or NULL to not make any).}
}
\value{
A list containing all the peak plot objects.
}
\description{
Visualize significant peaks
}
