% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_gene_locations}
\alias{validate_gene_locations}
\title{Reads and validates gene locations}
\usage{
validate_gene_locations(gene_locations)
}
\arguments{
\item{gene_locations}{File path or data.frame.}
}
\value{
validated gene locations data.frame
}
\description{
Reads and validates gene locations
}
