% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_arm_gistic.R
\name{do_arm_gistic}
\alias{do_arm_gistic}
\title{Apply BISCUT method to one chromosome arm amp/del tel/cent}
\usage{
do_arm_gistic(
  arm,
  direc,
  telcent,
  mode,
  ci,
  qval_thres,
  telcent_thres,
  breakpoint_file_dir,
  n,
  emp_bg,
  results_dir,
  abslocs,
  genelocs,
  seed = NULL
)
}
\description{
Apply BISCUT method to one chromosome arm amp/del tel/cent
}
\keyword{internal}
