% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_chr_coordinates}
\alias{validate_chr_coordinates}
\title{Reads and validates chromosome coordinates}
\usage{
validate_chr_coordinates(chromosome_coordinates)
}
\arguments{
\item{chromosome_coordinates}{File path or data.frame.}
}
\description{
Takes either a file path or a data.frame, validates the data.frame, and returns a version with
just the necessary columns.
}
